Writing Geometries
This script updates point geometries to another feature class based on a unique value.  If the two feature classes share a unique value, the geometry of the input feature will get written to the target feature.  This code is an version of code posted to ESRI's GeoNet forum by Jake Skinner.